+++ apps/backoffice-v2/src/lib/blocks/hooks/useDocumentBlocks/helpers/is-ubo-document.ts
@@ -1,0 +1,4 @@
+import { TDocument } from '@ballerine/common';
+
+export const isUboDocument = (uboDocuments: TDocument[], document: TDocument) =>
+  uboDocuments.some(uboDocument => uboDocument.id === document.id);