+++ apps/backoffice-v2/src/lib/blocks/hooks/useUbosDocuments/helpers/get-ubos-entity-ids-from-workflow.ts
@@ -1,0 +1,5 @@
+import { TWorkflowById } from '@/domains/workflows/fetchers';
+
+export const getUbosEntityIdsFromWorkflow = (workflow: TWorkflowById) => {
+  return workflow.childWorkflows?.map(childWorkflow => childWorkflow.entity.id) ?? [];
+};