+++ packages/ui/src/components/organisms/Form/Validator/validators/format/types.ts
@@ -1,3 +1,3 @@
 export interface IFormatValueValidatorParams {
-  format: 'email';
+  format: 'email' | 'phone';
 }