+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,28 @@
 # web-ui-sdk
 
+## 1.5.84
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.83
+
+## 1.5.83
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.82
+
+## 1.5.82
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.81
+
 ## 1.5.81
 
 ### Patch Changes