+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.105
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.105
+
+## 0.6.104
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.104
+
+## 0.6.103
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.103
+
+## 0.6.102
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.102
+
 ## 0.6.101
 
 ### Patch Changes